Extract BlockInner to its own component
